==> 0-d-process-replication-data.ipynb <==
import pandas as pd 
import numpy as np 
import matplotlib.pyplot as plt 
import datetime

## **REPLICATION DATA**

### **Barbera 2014: 'Birds of the Same Feather Tweet Together...' (Bayesian)**

elites_data = pd.read_csv('data/misc/Replication_data_Barbera_2014/elites-data.csv')
elites_data = elites_data[['screen_name','name', 'dw.nom.1', 'nameid', 'party', 'state', 'gender', 'idealPoint']]
print(f'{len(elites_data)} elites loaded')

results_elites_US = pd.read_csv('data/misc/Replication_data_Barbera_2014/results-elites-US.csv')
results_elites_US = results_elites_US.drop(['Unnamed: 0', 'phi.sd', 'alpha'], axis=1)
print(f'{len(results_elites_US)} results loaded')

barbera_2014_data = elites_data.merge(results_elites_US, on='screen_name', how='left')

# Filter just for congressmembers
barbera_2014_data_congressmembers = barbera_2014_data[(barbera_2014_data['phi'].notnull()) & (barbera_2014_data['dw.nom.1'].notnull())] \
                    .drop(['screen_name', 'name', 'idealPoint', 'state', 'party', 'gender'], axis=1) \
                    .rename(columns = {'phi' : 'Barbera1_score'})
print(f'{len(barbera_2014_data_congressmembers)} congressmembers loaded')

barbera_2014_data_congressmembers.sort_values(by='nameid', ascending=True).head(15)

#dw.nom1 is the first dimension DW-NOMINATE of congressmembers
#phi is estimated ideology score from logit/bayesian method (phi variable)

### **Barbera 2015: 'Tweeting from Left to Right...' (Correspondence)**

elites_data_2 = pd.read_csv('data/misc/Replication_data_Barbera_2015/elites-data.csv').rename(columns = {'id' : 'nameid'})
print(f'{len(elites_data_2)} elites loaded')
house_jackman_results = pd.read_csv('data/misc/Replication_data_Barbera_2015/house.csv')
print(f'{len(house_jackman_results)} house results loaded')
senate_jackman_results = pd.read_csv('data/misc/Replication_data_Barbera_2015/senate.csv')
print(f'{len(senate_jackman_results)} senate results loaded')
congresss_jackman_results = pd.concat([house_jackman_results, senate_jackman_results])[['nameid', 'idealPoint', 'gender']]
print(f'{len(congresss_jackman_results)} congress results loaded')
barbera_2015_data = elites_data_2.dropna(subset=['nameid']).merge(congresss_jackman_results, on='nameid', how='left') \
                    .drop(['thomas_id', 'twitter_id', 'type'], axis=1).rename(columns = {'idealPoint' : 'Jackman_score'})
barbera_2015_data.head(15)

## **ZHIXIANG DATA**

### **House_list_in_318-accounts.csv + Senate_list_in_318-accounts.csv (congressmember amongst the 318 elites)**

all_congressmembers = pd.read_csv('data/raw/all_congress_members.csv').drop(['DW-Nominate'], axis=1)[['nameid', 'twitter_name', 'congressperson']]

zhixiang_house = pd.read_csv('data/misc/Data_about_Opinion_Measurement_Zhixiang/2.Opinion results/opinion_house.csv')
zhixiang_senate = pd.read_csv('data/misc/Data_about_Opinion_Measurement_Zhixiang/2.Opinion results/opinion_senate.csv')
zhixiang_results = pd.concat([zhixiang_house, zhixiang_senate]).rename(columns = {'opinion measured' : 'Zhixiang_score'})

zhixiang_data = zhixiang_results.merge(all_congressmembers, on='twitter_name', how='left').dropna(subset=['nameid']).drop(['twitter_name'], axis=1)

full_characteristics_table = zhixiang_data.merge(barbera_2014_data_congressmembers, on='nameid', how='left') \
                                          .merge(barbera_2015_data, on='nameid', how='left') \
                                          .drop_duplicates(subset=['nameid']) \
                                          [['nameid', 'congressperson', 'first_name', 'last_name', 'gender','party', 'chamber', 'stdis','twitter_name', 'dw.nom.1', 'Barbera1_score', 'Jackman_score', 'Zhixiang_score']]
                                          

full_characteristics_table.to_csv('data/raw/congressmembers_characteristics.csv')

full_characteristics_table.head(10)

full_characteristics_table[full_characteristics_table['congressperson'] == 'senatorreid']
